Compile the CSS to a temporary file to avoid premature browser reloads
`browser-sync` attempts to reload the browser as soon as it detects a
change, but because our output file is so large the browser is reloading
the compiled CSS *before* the file has been completely written. This is
making it load an empty file, so all the styling is disappearing.

Instead, now we write to `.tmp/main.css` (which is the slow operation)
then we `mv .tmp/main.css out/css/main.css` which happens instantly.